% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headline-data.R
\name{headlines_data}
\alias{headlines_data}
\title{Data in list format used to make the experiment headlines}
\usage{
headlines_data(news_type = "hard")
}
\arguments{
\item{news_type}{Which news type to get, either \code{hard} (default) or \code{soft}.}
}
\description{
Data in list format used to make the experiment headlines.
}
\examples{
\dontrun{
 headlines_data_hard_news <- headlines_data("hard")
}
}
\keyword{headlines}
\keyword{prepublication}

