% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headlines-make.R
\name{headlines_make_from_data}
\alias{headlines_make_from_data}
\title{Make all variations of a headline from headline data}
\usage{
headlines_make_from_data(news_data)
}
\arguments{
\item{news_data}{Headline data in list format (see \code{headlines_data()}).}
}
\description{
Make all variations of a headline from headline data in list format (see \code{headlines_data()}). Returns a data frame with the headline, as well as variables identifying the integrated treatment component.
}
\examples{
\dontrun{
 hard_news_data <- headlines_data("hard")[[1]]
 hard_news <- headlines_make_from_data(hard_news_data[[1]])
 print(hard_news$headline[1])
}
}
\keyword{headlines}
\keyword{prepublication}

