% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headlines-save.R
\name{headlines_save}
\alias{headlines_save}
\title{Save headlines as file}
\usage{
headlines_save(folder = getwd(), csv_for_data_collector = TRUE,
  rdata_for_package = TRUE)
}
\arguments{
\item{folder}{Which folder to write the files. Defaults to working directory.}

\item{csv_for_data_collector}{Whether to write .csv-files (which we sendt to our data collector).}

\item{rdata_for_package}{Whether to write the .RData-file we included in this compendium (\code{data("headlines")}).}
}
\description{
Write headlines to file.
}
\examples{
\dontrun{
 headlines_save()
}
}
\keyword{headlines}
\keyword{prepublication}

