% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefielding-shiny.R
\name{launch_mockup}
\alias{launch_mockup}
\title{Launch experiment mockup used for piloting}
\usage{
launch_mockup(...)
}
\description{
Launch experiment mockup used for piloting.
}
\examples{
\dontrun{
 launch_mockup()
}
}
\keyword{piloting}
\keyword{prepublication}
\keyword{rshiny}

